requirements.txt
soundfile
numpy
pyserial